Game Lobby Functional, Customization started
Have Game Lobby Functionality mostly finished going to finish it with a
countdown when all the players have entered into the game. Started on
Customization, using 30 day free trail pro at the moment to work with
RTT on the GUI. (we need to get this finished quick lol). Finishing up
on Customization, Game Lobby then going to work on starting the game.